Renamed project to AdventureBookResolver
